Resolves #63 Room/views.py needs refactoring to put the email room invites into separate file
